GCP/PubSub/CloudFunctions/pubsub-publisher/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-pubsub==2.9.0